Move anything with frontier into a subfolder
