+++ setup.cfg
@@ -22,6 +22,7 @@ packages =
     ethereum/frontier/vm/precompiled_contracts
     ethereum/utils
     ethereum_optimized/
+    ethereum_optimized/frontier
 
 package_dir =
     =src
