+++ src/ethereum_optimized/frontier/state.py
