+++ src/ethereum_optimized/frontier/state_db.py
