+++ src/ethereum_optimized/frontier/trie.py
