+++ src/ethereum_optimized/frontier/trie_utils.py
